# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Website API Documentation",
        "description": "API documentation for AI Website backend service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/user/social/{driver}/redirect": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User with Social Redirect",
                "operationId": "user.social.redirect",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "Social Driver (google, facebook, github)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect URL generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\SocialRedirectResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Provider error"
                    }
                }
            }
        },
        "/api/v1/auth/user/social/{driver}/callback": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User with Social Callback",
                "operationId": "user.social.callback",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "Social Driver",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Social Login Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\AuthResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Authentication failed"
                    }
                }
            }
        },
...
        "/api/v1/videos/{video}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get a single video",
                "operationId": "videos.show",
                "parameters": [
                    {
                        "name": "video",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VideoResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update a video",
                "operationId": "videos.update",
                "parameters": [
                    {
                        "name": "video",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVideoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VideoResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "Delete a video",
                "operationId": "videos.destroy",
                "parameters": [
                    {
                        "name": "video",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpdateVideoRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "My Video",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "A sample video",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/video.mp4"
                    }
                },
                "type": "object"
            },
            "VideoResource": {
                "required": [
                    "id",
                    "code",
                    "driver",
                    "url",
                    "user_id",
                    "views",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "code": {
                        "type": "string",
                        "example": "aB3xK9mQ"
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "extension": {
                        "type": "string",
                        "example": "mp4",
                        "nullable": true
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "video/mp4",
                        "nullable": true
                    },
                    "driver": {
                        "type": "string",
                        "example": "direct"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 0
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "qualities": {
                        "type": "object",
                        "nullable": true
                    },
                    "views": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}